Change n and cap type

+++ parser/c_parser/c_parser.h
@@ -29,7 +29,7 @@ typedef enum {
 
 typedef struct cp_result {
     Symbol *syms;
-    size_t n, cap;
+    int n, cap;
     char *filename;
 } cp_result;
